move compare costs tab to individual module
